+++ CHANGELOG.md
@@ -10,6 +10,8 @@ Release date: ----
     - 2.4 tag and frame configuration (without reverse header support for tag)
     - added "TDRC" instead of deprecated "TYER"(2.3)/"TYE"(2.2): field with timestamp that has a fallback to string content.
     - new `RecordingDateTime` inside the `ID3Tag` class
+- Added "TDAT" frame support
+- Added "TIME" frame support
 
 #### Fixed
 - Refactoring `ID3TagEditor` class